+++ lib/bcdd/result.rb
@@ -2,32 +2,38 @@
 
 require_relative 'result/version'
 require_relative 'result/error'
-require_relative 'result/type'
 require_relative 'result/data'
 require_relative 'result/handler'
 require_relative 'result/failure'
 require_relative 'result/success'
 require_relative 'result/mixin'
+require_relative 'result/expectations'
 
 class BCDD::Result
   attr_accessor :unknown
 
-  attr_reader :_type, :value, :subject
+  attr_reader :subject, :data, :type_checker
 
   protected :subject
 
-  private :unknown, :unknown=
+  private :unknown, :unknown=, :type_checker
 
-  def initialize(type:, value:, subject: nil)
-    @_type = Type.new(type)
-    @value = value
+  def initialize(type:, value:, subject: nil, expectations: nil)
+    data = Data.new(name, type, value)
+
+    @type_checker = Expectations.evaluate(data, expectations)
     @subject = subject
+    @data = data
 
     self.unknown = true
   end
 
   def type
-    _type.to_sym
+    data.type
+  end
+
+  def value
+    data.value
   end
 
   def success?(_type = nil)
@@ -45,15 +51,15 @@ class BCDD::Result
   def on(*types, &block)
     raise Error::MissingTypeArgument if types.empty?
 
-    tap { known(block) if _type.in?(types, allow_empty: false) }
+    tap { known(block) if type_checker.allow?(types) }
   end
 
   def on_success(*types, &block)
-    tap { known(block) if success? && _type.in?(types, allow_empty: true) }
+    tap { known(block) if type_checker.allow_success?(types) && success? }
   end
 
   def on_failure(*types, &block)
-    tap { known(block) if failure? && _type.in?(types, allow_empty: true) }
+    tap { known(block) if type_checker.allow_failure?(types) && failure? }
   end
 
   def on_unknown
@@ -71,15 +77,11 @@ class BCDD::Result
   end
 
   def handle
-    handler = Handler.new(self)
+    handler = Handler.new(self, type_checker: type_checker)
 
     yield handler
 
     handler.send(:outcome)
-  end
-
-  def data
-    Data.new(self)
   end
 
   def ==(other)
@@ -108,7 +110,7 @@ class BCDD::Result
   private
 
   def name
-    raise Error::NotImplemented
+    :unknown
   end
 
   def known(block)