+++ lib/bcdd/result/success.rb
@@ -3,7 +3,7 @@
 class BCDD::Result
   class Success < self
     def success?(type = nil)
-      type.nil? || type == self.type
+      type.nil? || type_checker.allow_success?([type])
     end
 
     def failure?(_type = nil)