+++ test/bcdd/result/expectations/contract/for_types_and_values_test.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+require 'test_helper'
+
+class BCDD::Result::Expectations
+  class ContractForTypesAndValuesTest < Minitest::Test
+    test '#type?' do
+      contract = Contract::ForTypesAndValues.new(
+        ok: Object
+      )
+
+      assert contract.type?(:ok)
+      refute contract.type?(:yes)
+    end
+  end
+end